Move & modify remaining component files and utilities
