+++ girder/girder_large_image/web_client/package.json
@@ -24,7 +24,6 @@
         "jsonlint-mod": "^1.7.6",
         "sinon": "^2.1.0",
         "vue": "~2.6.14",
-        "vue-color": "^2.8.1",
         "vue-loader": "~15.9.8",
         "vue-template-compiler": "~2.6.14",
         "webpack": "^3",
