--- girder/girder_large_image/web_client/vue/utils/clamp.js
@@ -1,3 +1,0 @@
-export default function clamp(num, min, max) {
-    return Math.min(Math.max(num, min), max);
-}
