+++ setup.py
@@ -110,6 +110,7 @@ setup(
         'Pillow',
         'numpy',
         'ipyvue',
+        'aiohttp',
         'typing-extensions',
     ],
     extras_require=extraReqs,
